requirements.txt
supabase
langchain
langchain-community
pypdf
urllib3
requests_html